move adapter into its own file
